增加视频播放列表，增加ExoPlayer播放视频

+++ app/src/main/java/com/cyl/musiclake/utils/FormatUtil.kt
@@ -43,7 +43,7 @@ object FormatUtil {
      * 格式化播放次数
      *
      */
-    fun formatPlayCount(count: Int): String {
+    fun formatPlayCount(count: Long): String {
         return when {
             count < 10000 -> "$count"
             else -> {
